+++ src/microsui_core/sign.h
@@ -4,17 +4,21 @@
 #include <stdint.h>
 #include <stddef.h>
 
-int microsui_sign(uint8_t scheme, uint8_t sui_sig[97], const uint8_t* message, const size_t message_len, const uint8_t private_key[32]);
+int microsui_sign(uint8_t sui_sig_out[97], uint8_t scheme, const uint8_t* message, const size_t message_len, const uint8_t seed[32]);
 
-int microsui_sign_ed25519(uint8_t sui_sig[97], const uint8_t* message, const size_t message_len, const uint8_t private_key[32]);
+int microsui_sign_ed25519(uint8_t sui_sig_out[97], const uint8_t* message, const size_t message_len, const uint8_t seed[32]);
 
 int microsui_verify_signature_ed25519(uint8_t sui_sig[97], const uint8_t* message, const size_t message_len);
 
+int microsui_verify_signature_ed25519_with_public_key(uint8_t sui_sig[97], const uint8_t* message, const size_t message_len, uint8_t public_key[32]);
+
 int microsui_verify_signature(uint8_t sui_sig[97], const uint8_t* message, const size_t message_len);
+
+int microsui_verify_signature_with_public_key(uint8_t sui_sig[97], const uint8_t* message, const size_t message_len, uint8_t public_key[32]);
 
 /**
  * @deprecated  Use microsui_sign() or microsui_sign_ed25519() instead.
  */
-int microsui_sign_message(uint8_t sui_sig[97], const char* message_hex, const uint8_t private_key[32]);
+int microsui_sign_message(uint8_t sui_sig_out[97], const char* message_hex, const uint8_t seed[32]);
 
 #endif