+++ CHANGELOG.md
@@ -4,9 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## \[0.3.3] - 2026-04-01
+
+- Added new mod functions to Monocypher library (these functions are more efficient than the compact25519 ones).
+- Implemented new functions in `microsui_sign.c`, replacing the compact25519 ones.
+- Deleted the now-obsolete _compact25519 library_.
+- Refactor: rename output parameters for consistency in core functions.
+- Refactor: `private/secret key` in bytes now is renamed to `seed` for avoid imprecisions.
+- Added new Signature verification/check with public key function.
+- Added new examples (Signature and Benchmark).
+- Updated Makefiles.
+- Added GitHub `make` workflow.
+- General improvements and minor bugs fixed.
+
 ## \[0.3.2] - 2026-01-27
 
-This release introduces ***Signature Verification*** support and improves developer experience with clearer examples.
+This release introduces **_Signature Verification_** support and improves developer experience with clearer examples.
 
 ### Added